requirements.txt
web3==6.0.0
psycopg2-binary==2.9.6
SQLAlchemy==2.0.1
python-dotenv==1.0.0
requests==2.28.2
python-telegram-bot==20.0
gunicorn==20.1.0
prometheus_client==0.16.0
